pyproject.toml
[tool.poetry]
name = "vailtools"
version = "0.4.0"
license="MIT"
description = "Components and tools for building, training, and evaluating artificial neural networks in Python."
authors = ["Colin Van Oort <cvanoort156@gmail.com>", "VaiL <vail.csds@gmail.com>"]
readme = "README.md"
repository = "https://gitlab.com/vail-uvm/vail-tools"
keywords = ["TensorFlow"]
classifiers=[
    # Trove classifiers
    # Full list: https://pypi.python.org/pypi?%3Aaction=list_classifiers
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: Implementation :: CPython",
]
include = [
    "LICENSE.txt",
]

[tool.poetry.dependencies]
python = "^3.7"
# Optional to prevent poetry from mangling a conda env
numpy = {version = "*", optional = true}
scikit-image = "*"
# Optional to prevent poetry from mangling a conda env
tensorflow-gpu = {version = "*", optional = true}

[tool.poetry.dev-dependencies]
bandit = "*"
black = {version = "*", allow-prereleases = true}
isort = "*"
m2r = "*"
pytest = "*"
pytest-cov = "*"
# Keep sphinx >3 until M2R updates
sphinx = "^2"
vulture = "*"

# isort config taken from https://github.com/psf/black
[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
